Refactor: dto -> inner static class로 리팩토링
See Also: #23, #25
