--- module-api/src/main/java/com/zoopi/controller/member/request/AuthCodeCheckRequest.java
@@ -1,40 +1,0 @@
-package com.zoopi.controller.member.request;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
-
-import io.swagger.annotations.ApiModelProperty;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import com.zoopi.domain.phoneauthentication.entity.PhoneAuthenticationType;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AuthCodeCheckRequest {
-
-	@NotNull
-	@ApiModelProperty(value = "인증 타입", required = true, example = "SIGN_UP")
-	private PhoneAuthenticationType type;
-
-	@NotBlank
-	@Size(max = 50)
-	@ApiModelProperty(value = "인증 키", required = true, example = "2a6ec3d7-7147-4cdc-a6c5-fc5d937cfe76")
-	private String authenticationKey;
-
-	@NotNull
-	@Pattern(regexp = "^01(?:0|1|[6-9])(?:\\d{3}|\\d{4})\\d{4}$")
-	@ApiModelProperty(value = "휴대폰 번호", required = true, example = "01012345678")
-	private String phone;
-
-	@NotBlank
-	@Size(min = 6, max = 6)
-	@ApiModelProperty(value = "인증 코드", required = true, example = "012345")
-	private String authenticationCode;
-
-}
