+++ module-domain/src/main/java/com/zoopi/domain/member/dto/SigninResponse.java
@@ -2,14 +2,11 @@ package com.zoopi.domain.member.dto;
 
 import com.zoopi.domain.phoneauthentication.dto.JwtDto;
 
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class SigninResponse {
 
 	private JwtDto jwt;
